feat: prepare separate views
