+++ src/view/list-empty.js
@@ -1,0 +1,5 @@
+export const createEmptyListTemplate = () => (
+  `<p class="trip-events__msg">
+    Click New Event to create your first point
+  </p>`
+);
