+++ src/view/loading.js
@@ -1,0 +1,5 @@
+export const createLoadingTemplate = () => (
+  `<p class="trip-events__msg">
+    Loading...
+  </p>`
+);
