+++ src/view/stats.js
@@ -1,0 +1,13 @@
+export const createStatsTemplate = () => (`
+  <div class="statistics__item">
+    <canvas class="statistics__chart" id="money" width="900"></canvas>
+  </div>
+
+  <div class="statistics__item">
+    <canvas class="statistics__chart" id="type" width="900"></canvas>
+  </div>
+
+  <div class="statistics__item">
+    <canvas class="statistics__chart" id="time-spend" width="900"></canvas>
+  </div>
+`);
